Fixes to properly point to the moved components
